SERVER-43239 Fixed bug causing numCatchUpOps in repSetGetStatus to be incorrect

+++ src/mongo/db/repl/replication_coordinator.h
@@ -888,7 +888,7 @@ public:
      * Increment the counter for the number of ops applied during catchup if the node is in catchup
      * mode.
      */
-    virtual void incrementNumCatchUpOpsIfCatchingUp(int numOps) = 0;
+    virtual void incrementNumCatchUpOpsIfCatchingUp(long numOps) = 0;
 
     /**
      * Signals that drop pending collections have been removed from storage.
